Optimise get_mapper to avoid repeat lookups #144
get_mapper was being around 20k times in the benchmark test suite.  This was due to
mapper's being re-instantiated for each pipeline.  Avoiding calling get_mapper inside of
loops and repeat calls for the same field (IE Nested) has brough this number down to 4000
calls for the entire suite. This has yielded another 25% decrease in runtime.