Raise error when trying to serialize None
If you pass data instead of obj to a mapper and then try to serialise it, weird errors are likely to occur.

Serialise was calling the get_obj() method which should only be called on marshal, for serialise returning self.obj is enough - there is no benefit in instantiating a blank object if it's None. Instead detect if None and raise a helpful error.