Add tests and clarify docs around required/allow_none
Contrary to what we had assumed, allow_none is actually only relevant when a field is *not* required. Whilst it's a bit poorly named, there is in fact no problem with what it does and it's reason for existence.

Updated docs to make it more clear what the purpose of it is.